+++ include/Vtk.h
@@ -5,13 +5,13 @@
   #include<Mystdbool.h>
   #include<Mesh.h>
   void headVtk(char *,bool, FILE *);
-  void writeVtkCoor(double *x,long nnode     ,int ndm,bool cod, FILE *f);
-  void writeVtkCell(int *el  ,short int *nen , short int *type
-                   ,long numel ,short int maxno,bool cod
+  void writeVtkCoor(double *x,INT nnode   ,int ndm,bool cod, FILE *f);
+  void writeVtkCell(int *el   ,short *nen ,short *type
+                   ,INT numel ,short maxno,bool cod
                    ,FILE *f);
-  void writeVtkCellProp(int *iprop,double *dprop,long int numel     
-                       ,int gdl       ,char *s      ,bool cod1 
-                       ,short int cod2,FILE *f);
+  void writeVtkCellProp(int *iprop,double *dprop,INT  numel     
+                       ,int gdl   ,char *s      ,bool cod1 
+                       ,short cod2,FILE *f);
   void writeVtkNodeProp(int *iprop,double *dprop,short cod1,short cod2
                        ,int nnode ,short ndf    ,char *s   ,bool cod
                        ,FILE *f);
