+++ include/WriteVtk.h
@@ -9,11 +9,11 @@
   #include<Mesh.h>
   #include<Memoria.h>
   #include<File.h>
-  void wGeoVtk(Memoria *m        ,double *x      ,long int *el 
-              ,short int *mat    ,short int *nen ,short int *type
-              ,short int *faceRt1,double *faceSt1,long int nnode 
-              ,long int numel    ,long int ndm   ,short int maxno
-              ,short int *ndfT   ,char *nameOut  ,bool iws      
+  void wGeoVtk(Memoria *m    ,double *x      ,INT *el 
+              ,short *mat    ,short int *nen ,short *type
+              ,short *faceRt1,double *faceSt1,INT nnode 
+              ,INT numel     ,INT ndm        ,short maxno
+              ,short *ndfT   ,char *nameOut  ,bool iws      
               ,FILE *f); 
   
 #endif/*_WRITE_VTK_H*/
